+++ .eslintignore
@@ -7,8 +7,7 @@ coverage
 
 *.config.js
 
-cypress
+cypress/
 
 **/__tests__/
-**/plugin/**/*.js
-**/plugins/
+**/plugins/**/*.js
