+++ src/config/bridge/ethereum-darwinia.ts
@@ -1,4 +1,4 @@
-import { darwiniaConfig, ethereumConfig } from '..';
+import { darwiniaConfig, ethereumConfig } from '../network';
 import { Bridge, EthereumDarwiniaBridgeConfig } from '../../model';
 import { EVOLUTION_DOMAIN } from '../api';
