+++ src/utils/network/graph.ts
@@ -1,6 +1,6 @@
 import { isEqual } from 'lodash';
+import { BRIDGES } from '../../config';
 import { Arrival, Bridge, BridgeConfig, Departure } from '../../model';
-import { BRIDGES } from '../bridge';
 
 const isPro = process.env.REACT_APP_HOST_TYPE === 'prod';
